python-api/requirements.txt
fastapi
uvicorn
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.16
langsmith
chromadb
pandas
pydantic
python-dotenv